+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
   - Export `DBSchemaError` and `DBFormatError` in main module in order to
     facilitate their import from other software that use RacksDB as an external
     library.
+  - Add optional `path` property to `DBLoader` class.
 - cli: Add `-c, --coordinates` and `--coordinates-format` arguments on
   `racksdb draw` sub-command to generate coordinate file in either JSON or YAML
   format with coordinates of racks and equipment in the image along with the